+++ test/test_fx_experimental.py
@@ -4,8 +4,8 @@ from typing import Dict
 from torch.fx.symbolic_trace import symbolic_trace
 from torch.fx.graph_module import GraphModule
 from torch.fx.node import Node
-from torch.fx.experimental import GraphManipulation
-from torch.fx.experimental.Partitioner import Partitioner
+from torch.fx.experimental import graph_manipulation
+from torch.fx.experimental.accelerator_partitioner import Partitioner
 from torch.fx.experimental.rewriter import RewritingTracer
 from torch.testing._internal.common_utils import run_tests
 from torch.testing._internal.jit_utils import JitTestCase
@@ -55,7 +55,7 @@ class TestFXExperimental(JitTestCase):
         a = torch.rand(4)
         b = torch.rand(4)
         c = torch.rand(3, 3, 2, 2)
-        GraphManipulation.get_size_of_all_nodes(traced, [a, b, c])
+        graph_manipulation.get_size_of_all_nodes(traced, [a, b, c])
 
         partitioner = Partitioner()
         devices = [Device("dev_0", 5000, 0), Device("dev_1", 125, 1)]
@@ -76,8 +76,8 @@ class TestFXExperimental(JitTestCase):
             node.shape = a.shape
             node.dtype = a.dtype
 
-        agm1 = GraphManipulation.AcceleratedGraphModule(traced)
-        agm2 = GraphManipulation.AcceleratedGraphModule(module_with_submodules)
+        agm1 = graph_manipulation.AcceleratedGraphModule(traced)
+        agm2 = graph_manipulation.AcceleratedGraphModule(module_with_submodules)
         assert len(agm1.weights) == 4
         assert len(agm2.weights) == 4
         assert len(agm1.serialized_graph["nodes"]) == 10
@@ -108,8 +108,8 @@ class TestFXExperimental(JitTestCase):
         q_tensor_channel = torch.quantize_per_channel(
             x, torch.tensor([0.1, 0.01]), torch.tensor([10, 0]), 0, torch.quint8
         )
-        result = GraphManipulation.serialize_tensor_quantization(q_tensor)
-        result2 = GraphManipulation.serialize_tensor_quantization(q_tensor_channel)
+        result = graph_manipulation.serialize_tensor_quantization(q_tensor)
+        result2 = graph_manipulation.serialize_tensor_quantization(q_tensor_channel)
         assert result["q_scheme"] == "torch.per_tensor_affine"
         assert result["q_scale"] == 1.0
         assert result2["q_scheme"] == "torch.per_channel_affine"
@@ -124,7 +124,7 @@ class TestFXExperimental(JitTestCase):
         traced = symbolic_trace(m)
         a = torch.rand(1)
         b = torch.rand(1)
-        GraphManipulation.get_size_of_all_nodes(traced, [a, b])
+        graph_manipulation.get_size_of_all_nodes(traced, [a, b])
         partitioner = Partitioner()
         devices = [
             Device("dev_0", 125, 0),
@@ -155,7 +155,7 @@ class TestFXExperimental(JitTestCase):
         traced = symbolic_trace(m)
         a = torch.rand(4)
         b = torch.rand(4)
-        GraphManipulation.get_size_of_all_nodes(traced, [a, b])
+        graph_manipulation.get_size_of_all_nodes(traced, [a, b])
         partitioner = Partitioner()
         devices = [
             Device("dev_0", 125, 0),
@@ -187,7 +187,7 @@ class TestFXExperimental(JitTestCase):
         m = TestModule()
         traced = symbolic_trace(m)
         a = torch.rand(4)
-        GraphManipulation.get_size_of_all_nodes(traced, [a])
+        graph_manipulation.get_size_of_all_nodes(traced, [a])
         partitioner = Partitioner()
         devices = [Device("dev_0", 120, 0), Device("dev_1", 160, 1)]
         partitioner_config = PartitionerConfig(devices, is_sparse_nn=False)
@@ -249,7 +249,7 @@ class TestFXExperimental(JitTestCase):
         b = torch.randint(10, (8,))
         offset = torch.randint(1, (2,))
         traced = symbolic_trace(m)
-        GraphManipulation.get_size_of_all_nodes(traced, [a, b, offset])
+        graph_manipulation.get_size_of_all_nodes(traced, [a, b, offset])
         devices = [
             Device("dev_0", 33000000, 0),
             Device("dev_1", 33000000, 1),
@@ -297,7 +297,7 @@ class TestFXExperimental(JitTestCase):
         m = TestModule()
         traced = symbolic_trace(m)
         a = torch.rand(4)
-        GraphManipulation.get_size_of_all_nodes(traced, [a])
+        graph_manipulation.get_size_of_all_nodes(traced, [a])
         node_to_latency_mapping = get_node_to_latency_mapping(traced)
         devices = [Device("dev_0", 200, 0), Device("dev_1", 200, 1)]
         partitioner = Partitioner()
@@ -348,7 +348,7 @@ class TestFXExperimental(JitTestCase):
         m = MyModule()
         traced = symbolic_trace(m)
         a = torch.rand(4)
-        GraphManipulation.get_size_of_all_nodes(traced, [a])
+        graph_manipulation.get_size_of_all_nodes(traced, [a])
         devices = [
             Device('dev_0', 125, 0),
             Device('dev_1', 125, 1),
@@ -376,6 +376,24 @@ class TestFXExperimental(JitTestCase):
             partitioner_config.transfer_rate_bytes_per_sec
         )
         assert critical_path_latency_sec == 160.
+
+        def test_replace_target_nodes_with(self):
+            class testModule(torch.nn.Module):
+                def forward(self, a, b):
+                    return a + b
+            m = testModule()
+            traced = symbolic_trace(m)
+            input1 = torch.randn(1)
+            input2 = torch.randn(1)
+            assert (input1 + input2) == traced(input1, input2)
+            graph_manipulation.replace_target_nodes_with(
+                fx_module=traced,
+                old_op="call_function",
+                old_target=operator.add,
+                new_op="call_function",
+                new_target=operator.mul,
+            )
+            assert (input1 * input2) == traced(input1, input2)
 
     def test_call_to_assert_no_msg(self):
         class M(torch.nn.Module):
